+++ src/includes/paradise.h
@@ -19,3 +19,4 @@ WRITE_HANDLER( paradise_palbank_w );
 
 VIDEO_START( paradise );
 VIDEO_UPDATE( paradise );
+VIDEO_UPDATE( torus );